+++ ringo-domain/src/main/java/no/sr/ringo/message/PeppolMessageNotFoundException.java
@@ -5,7 +5,7 @@ package no.sr.ringo.message;
  * @author Steinar Overbeck Cook steinar@sendregning.no
  */
 public class PeppolMessageNotFoundException extends IllegalArgumentException {
-    public PeppolMessageNotFoundException(Integer messageNo) {
+    public PeppolMessageNotFoundException(Long messageNo) {
         super("PEPPOL message number " + messageNo + " not found");
     }
 }
